third_party/blink/web_tests/svg/dynamic-updates/SVGForeignObjectElement-svgdom-x-prop-expected.txt

SVG 1.1 dynamic update tests
Test passed

Tests dynamic updates of the 'x' property of the SVGForeignObjectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foreignObjectElement.x.baseVal.value is -100
PASS foreignObjectElement.x.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
